renaming folder and moving component to EventComponent folder
